requirements.txt
Flask==0.12.1
Flask-Bootstrap==3.3.7.1
Pillow==4.1.0
requests==2.13.0
flask-table==0.5.0